+++ src/main/java/org/uom/lefterisxris/codetour/tours/service/AppSettingsConfigurable.java
@@ -4,6 +4,7 @@ import com.intellij.openapi.options.Configurable;
 import org.jetbrains.annotations.Nls;
 import org.jetbrains.annotations.Nullable;
 import org.uom.lefterisxris.codetour.tours.ui.AppSettingsComponent;
+import org.uom.lefterisxris.codetour.tours.ui.CodeTourNotifier;
 
 import javax.swing.*;
 import java.util.Optional;
@@ -62,6 +63,11 @@ public class AppSettingsConfigurable implements Configurable {
       settingsComponent.setOnboardingAssistant(settings.isOnboardingAssistant());
       settingsComponent.setSortOption(settings.getSortOption());
       settingsComponent.setSortDirection(settings.getSortDirection());
+      //TODO: This should be done automatically, instead of just prompting user
+
+      // Notify user to reload Settings
+      CodeTourNotifier.warn(null, "CodeTour User Settings Changed: " +
+            "Reload Tours from the bottom right button in Tool Pane Window");
    }
 
    @Override